requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
Flask-SocketIO==4.3.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
pyserial==3.5
python-engineio==3.14.2
python-socketio==4.6.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1